Move tests to integ space to include in-memory db
