+++ indexer/tasks/Cargo.toml
@@ -27,6 +27,10 @@ serde = { version = "1.0", features = ["derive"] }
 tracing-subscriber = "0.3.9"
 tracing = "0.1.31"
 
+[dev-dependencies]
+rand = "0.8.5"
+sea-orm-macros = "0.8.0"
+
 [features]
 default = ["build_rust_task"]
 build_rust_task = []
